chore: move deployment resources interfaces
