+++ packages/amplify-category-api/src/graphql-transformer/transform-graphql-schema-v2.ts
@@ -1,12 +1,15 @@
 import {
-  DeploymentResources,
   GraphQLTransform,
   OverrideConfig,
   ResolverConfig,
-  Template,
   TransformerProjectConfig,
 } from '@aws-amplify/graphql-transformer-core';
-import { AppSyncAuthConfiguration, TransformerPluginProvider } from '@aws-amplify/graphql-transformer-interfaces';
+import {
+  AppSyncAuthConfiguration,
+  DeploymentResources,
+  Template,
+  TransformerPluginProvider,
+} from '@aws-amplify/graphql-transformer-interfaces';
 import {
   $TSContext,
   AmplifyCategories,
